completeTodos is stored in a module
