+++ src/index.js
@@ -2,12 +2,14 @@ import './style.css';
 import todos from './modules/todos.js';
 import storageAvailable from './modules/storageAvailable.js';
 import { clearBtn, todoInput } from './modules/htmlElements.js';
+import completeToDo from './modules/completedCheck.js';
 
 if (storageAvailable('localStorage')) {
   todos.checkStorage();
 
   todos.showTodos();
 
+  // Remove a To-Do
   for (let i = 0; i < todos.todos.length; i += 1) {
     const removeButtons = document.querySelectorAll('.remove-button');
     removeButtons[i].addEventListener('click', () => {
@@ -15,18 +17,10 @@ if (storageAvailable('localStorage')) {
     });
   }
 
-  for (let i = 0; i < todos.todos.length; i += 1) {
-    const checkboxes = document.querySelectorAll('.checkbox');
-    checkboxes[i].addEventListener('change', (e) => {
-      if (e.target.checked) {
-        todos.complete(i, true);
-      } else {
-        todos.complete(i, false);
-      }
-      todos.save();
-    });
-  }
+  // Mark a To-Do as completed
+  completeToDo(todos);
 
+  // Edit a To-Do
   for (let i = 0; i < todos.todos.length; i += 1) {
     const todoTexts = document.querySelectorAll('.todo-text');
     todoTexts[i].addEventListener('keydown', (event) => {
@@ -40,6 +34,7 @@ if (storageAvailable('localStorage')) {
     });
   }
 
+  // Add a To-Do
   todoInput.addEventListener('keyup', (event) => {
     if (todoInput.value !== '') {
       if (event.keyCode === 13) {
@@ -55,6 +50,7 @@ if (storageAvailable('localStorage')) {
     }
   });
 
+  // Clean completed To-Dos
   clearBtn.addEventListener('click', () => {
     todos.clearCompleted();
     todos.save();
